move files around to avoid cycles
